+++ components/micro-gateway-cli/src/main/java/org/wso2/apimgt/gateway/cli/cmd/GatewayCliCommands.java
@@ -24,4 +24,5 @@ public class GatewayCliCommands {
     public static final String SETUP = "setup";
     public static final String BUILD = "build";
     public static final String RUN = "run";
+    public static final String RESET = "reset";
 }
