+++ components/micro-gateway-cli/src/main/java/org/wso2/apimgt/gateway/cli/cmd/Main.java
@@ -33,6 +33,7 @@ import org.wso2.apimgt.gateway.cli.config.TOMLConfigParser;
 import org.wso2.apimgt.gateway.cli.constants.GatewayCliConstants;
 import org.wso2.apimgt.gateway.cli.constants.RESTServiceConstants;
 import org.wso2.apimgt.gateway.cli.exception.BallerinaServiceGenException;
+import org.wso2.apimgt.gateway.cli.exception.CLIRuntimeException;
 import org.wso2.apimgt.gateway.cli.exception.CliLauncherException;
 import org.wso2.apimgt.gateway.cli.exception.ConfigParserException;
 import org.wso2.apimgt.gateway.cli.exception.HashingException;
@@ -83,54 +84,8 @@ public class Main {
         } catch (CliLauncherException e) {
             outStream.println(e.getMessages());
             Runtime.getRuntime().exit(1);
-        }
-    }
-
-    private static void init(String overrideProjectRootPath, String label, String configPath) {
-        try {
-            String projectRoot;
-            if (StringUtils.isBlank(overrideProjectRootPath)) {
-                String storedProjectRoot = GatewayCmdUtils.getStoredWorkspaceLocation();
-                if (StringUtils.isBlank(storedProjectRoot)) {
-                    outStream.println("Stored workspace path not available. "
-                            + "You need to specify --path <path to generate resources>");
-                    Runtime.getRuntime().exit(1);
-                }
-                projectRoot = storedProjectRoot;
-            } else {
-                projectRoot = overrideProjectRootPath;
-                GatewayCmdUtils.storeWorkspaceLocation(projectRoot);
-                GatewayCmdUtils.createMainProjectStructure(projectRoot);
-                GatewayCmdUtils.createLabelProjectStructure(projectRoot, label);
-                GatewayCmdUtils.createLabelConfig(projectRoot, label);
-            }
-
-            //user can define different label time to time. So need to create irrespective path provided or not.
-            GatewayCmdUtils.createLabelProjectStructure(projectRoot, label);
-
-            Path configurationFile = Paths.get(configPath);
-            if (Files.exists(configurationFile)) {
-                Config config = TOMLConfigParser.parse(configPath, Config.class);
-                GatewayCmdUtils.setConfig(config);
-            } else {
-                outStream.println("Config: " + configPath + " Not found.");
-                Runtime.getRuntime().exit(1);
-            }
-
-            String labelConfigPath = GatewayCmdUtils.getLabelConfigLocation(projectRoot, label);
-            ContainerConfig containerConfig = TOMLConfigParser.parse(labelConfigPath, ContainerConfig.class);
-            GatewayCmdUtils.setContainerConfig(containerConfig);
-
-            CodeGenerationContext codeGenerationContext = new CodeGenerationContext();
-            codeGenerationContext.setLabel(label);
-            GatewayCmdUtils.setCodeGenerationContext(codeGenerationContext);
-        } catch (ConfigParserException e) {
-            outStream.println(
-                    "Error while parsing the config" + (e.getCause() != null ? ": " + e.getCause().getMessage() : ""));
-            Runtime.getRuntime().exit(1);
-        } catch (IOException e) {
-            e.printStackTrace();
-            outStream.println("Error while processing files:" + e.getMessage());
+        } catch (CLIRuntimeException e) {
+            outStream.println(e.getMessage());
             Runtime.getRuntime().exit(1);
         }
     }
@@ -162,6 +117,10 @@ public class Main {
             RunCmd runCmd = new RunCmd();
             cmdParser.addCommand(GatewayCliCommands.RUN, runCmd);
             runCmd.setParentCmdParser(cmdParser);
+
+            ResetCmd resetCmd = new ResetCmd();
+            cmdParser.addCommand(GatewayCliCommands.RESET, resetCmd);
+            resetCmd.setParentCmdParser(cmdParser);
 
             cmdParser.setProgramName(MICRO_GW);
             cmdParser.parse(args);
@@ -199,585 +158,5 @@ public class Main {
         }
     }
 
-    private static void printUsageInfo(String commandName) {
-        String usageInfo = GatewayLauncherCmd.getCommandUsageInfo(commandName);
-        outStream.println(usageInfo);
-    }
 
-    /**
-     * This class represents the "help" command and it holds arguments and flags specified by the user.
-     */
-    @Parameters(commandNames = "help", commandDescription = "print usage information")
-    private static class HelpCmd implements GatewayLauncherCmd {
-
-        @Parameter(description = "Command name")
-        private List<String> helpCommands;
-
-        @Parameter(names = "--java.debug", hidden = true)
-        private String javaDebugPort;
-
-        private JCommander parentCmdParser;
-
-        public void execute() {
-            if (helpCommands == null) {
-                printUsageInfo(GatewayCliCommands.HELP);
-                return;
-
-            } else if (helpCommands.size() > 1) {
-                throw GatewayCmdUtils.createUsageException("too many arguments given");
-            }
-
-            String userCommand = helpCommands.get(0);
-            if (parentCmdParser.getCommands().get(userCommand) == null) {
-                throw GatewayCmdUtils.createUsageException("unknown help topic `" + userCommand + "`");
-            }
-
-            String commandUsageInfo = GatewayLauncherCmd.getCommandUsageInfo(userCommand);
-            outStream.println(commandUsageInfo);
-        }
-
-        @Override
-        public String getName() {
-            return GatewayCliCommands.HELP;
-        }
-
-        @Override
-        public void setParentCmdParser(JCommander parentCmdParser) {
-            this.parentCmdParser = parentCmdParser;
-        }
-
-    }
-
-    /**
-     * This class represents the "setup" command and it holds arguments and flags specified by the user.
-     */
-    @Parameters(commandNames = "setup", commandDescription = "setup information")
-    private static class SetupCmd implements GatewayLauncherCmd {
-
-        @Parameter(names = "--java.debug", hidden = true)
-        private String javaDebugPort;
-
-        @Parameter(names = { "-u", "--user" }, hidden = true)
-        private String username;
-
-        @Parameter(names = { "-p", "--password" }, hidden = true)
-        private String password;
-
-        @Parameter(names = { "-l", "--label" }, hidden = true)
-        private String label;
-
-        @Parameter(names = { "--reset" }, hidden = true)
-        private boolean reset;
-
-        @Parameter(names = { "--path" }, hidden = true)
-        private String path;
-
-        @Parameter(names = { "--base-url" }, hidden = true)
-        private String baseUrl;
-
-        @Parameter(names = { "-t", "--trustStore" }, hidden = true)
-        private String trustStoreLocation;
-
-        @Parameter(names = { "-s", "--trustStorePass" }, hidden = true)
-        private String trustStorePassword;
-
-        @Parameter(names = { "-n", "--project" }, hidden = true)
-        private String projectName;
-
-        @Parameter(names = { "-c", "--config" }, hidden = true)
-        private String configPath;
-
-        private String publisherEndpoint;
-        private String adminEndpoint;
-        private String registrationEndpoint;
-        private String tokenEndpoint;
-        private String clientID;
-        private String clientSecret;
-
-        public void execute() {
-            //initialize CLI with the provided path. First time the cli runs it is a must to provide this. Once it is
-            // provided, it is stored in <CLI_HOME>/temp/workspace.txt. In next runs, no need to provide the path and
-            // path is taken from above file.
-            if (StringUtils.isEmpty(label)) {
-                outStream.println("Label can't be empty. You need to specify -l <label name>");
-                return;
-            }
-
-            if (StringUtils.isEmpty(projectName)) {
-                projectName = label;
-            }
-
-            if (StringUtils.isEmpty(configPath)) {
-                configPath = GatewayCmdUtils.getMainConfigLocation();
-            }
-
-            init(path, projectName, configPath);
-
-            Config config = GatewayCmdUtils.getConfig();
-            String projectRoot = StringUtils.EMPTY;
-            try {
-                projectRoot = GatewayCmdUtils.getStoredWorkspaceLocation();
-            } catch (IOException e) {
-                e.printStackTrace();
-                outStream.println("Stored workspace path not available. "
-                        + "You need to specify --path <path to generate resources>");
-                Runtime.getRuntime().exit(1);
-            }
-
-            if (reset) {
-                config.getToken().setClientId(StringUtils.EMPTY);
-                config.getToken().setClientSecret(StringUtils.EMPTY);
-                config.getToken().setUsername(StringUtils.EMPTY);
-                try {
-                    TOMLConfigParser.write(GatewayCmdUtils.getMainConfigDirPath(projectRoot), config);
-                } catch (ConfigParserException e) {
-                    e.printStackTrace();
-                    outStream.println("Error occurred while writing to the config file");
-                    Runtime.getRuntime().exit(1);
-                }
-            }
-
-            boolean isOverwriteRequired = false;
-
-            //Setup username
-            //If config is there
-            if (config != null) {
-                //get previous user
-                String configuredUser = config.getToken().getUsername();
-                // check username is also provided
-                if (!StringUtils.isEmpty(username)) {
-                    // if provided user is different user it and mark as config is need to changed
-                    if (!username.equalsIgnoreCase(configuredUser)) {
-                        username = configuredUser;
-                        isOverwriteRequired = true;
-                    }
-                    // if username is not provided, get the previous user
-                } else {
-                    if (StringUtils.isEmpty(configuredUser)) {
-                        isOverwriteRequired = true;
-                        if ((username = promptForTextInput("Enter Username: ")).trim().isEmpty()) {
-                            throw GatewayCmdUtils.createUsageException("Micro gateway setup failed: empty username.");
-                        }
-                    } else {
-                        username = configuredUser;
-                    }
-                }
-            } else {
-                isOverwriteRequired = true;
-                // If there is no config file, prompt user to enter
-                if (StringUtils.isEmpty(username)) {
-                    if ((username = promptForTextInput("Enter Username: ")).trim().isEmpty()) {
-                        if ((username = promptForTextInput("Enter Username: ")).trim().isEmpty()) {
-                            throw GatewayCmdUtils.createUsageException("Micro gateway setup failed: empty username.");
-                        }
-                    }
-                }
-            }
-
-            //Setup password
-            if (StringUtils.isEmpty(password)) {
-                if ((password = promptForPasswordInput("Enter Password for " + username + ": ")).trim().isEmpty()) {
-                    if (StringUtils.isEmpty(password)) {
-                        password = promptForPasswordInput(
-                                "Password can't be empty; enter password for " + username + ": ");
-                        if (password.trim().isEmpty()) {
-                            throw GatewayCmdUtils.createUsageException("Micro gateway setup failed: empty password.");
-                        }
-                    }
-                }
-            }
-
-            //Setup urls
-            //If config is there
-            if (config != null) {
-                //get previous publisher endpoint
-                String existingPublisherEndpoint = config.getToken().getPublisherEndpoint();
-                // check baseUrl is also provided
-                if (!StringUtils.isEmpty(baseUrl)) {
-                    // if provided baseUrl is different, use it and mark as config is need to changed
-                    String newPublisherEndpoint;
-                    try {
-                        newPublisherEndpoint = new URL(new URL(baseUrl), RESTServiceConstants.PUB_RESOURCE_PATH)
-                                .toString();
-                    } catch (MalformedURLException e) {
-                        outStream.println("Existing publisher endpoint is malformed. Please reset and setup again.");
-                        throw new RuntimeException(
-                                "Existing publisher endpoint is malformed. Please reset and setup again.");
-                    }
-                    if (!newPublisherEndpoint.toString().equalsIgnoreCase(existingPublisherEndpoint)) {
-                        // If baseUrl is different derive urls again
-                        isOverwriteRequired = true;
-                    }
-                    populateHosts(baseUrl);
-                } else {
-                    //If same url is provided, use the existing config
-                    publisherEndpoint = config.getToken().getPublisherEndpoint();
-                    adminEndpoint = config.getToken().getAdminEndpoint();
-                    registrationEndpoint = config.getToken().getRegistrationEndpoint();
-                    tokenEndpoint = config.getToken().getTokenEndpoint();
-                    if (StringUtils.isEmpty(publisherEndpoint) || StringUtils.isEmpty(adminEndpoint) || StringUtils
-                            .isEmpty(registrationEndpoint) || StringUtils.isEmpty(tokenEndpoint)) {
-                        isOverwriteRequired = true;
-                        if ((baseUrl = promptForTextInput(
-                                "Enter APIM base URL [" + RESTServiceConstants.DEFAULT_HOST + "]: ")).trim()
-                                .isEmpty()) {
-                            baseUrl = RESTServiceConstants.DEFAULT_HOST;
-                            populateHosts(baseUrl);
-                        }
-                    }
-                }
-            } else {
-                // If there is no config file, prompt user to enter baseUrl
-                if (StringUtils.isEmpty(baseUrl)) {
-                    isOverwriteRequired = true;
-                    if ((baseUrl = promptForTextInput(
-                            "Enter APIM base URL [" + RESTServiceConstants.DEFAULT_HOST + "]: ")).trim().isEmpty()) {
-                        baseUrl = RESTServiceConstants.DEFAULT_HOST;
-                    }
-                }
-                // derive urls with given url
-                populateHosts(baseUrl);
-            }
-
-            //configure trust store
-            //If config is there
-            if (config != null) {
-                //get previous trust store
-                String configuredTrustStore = config.getToken().getTrustStoreLocation();
-                // check trust store is also provided
-                if (!StringUtils.isEmpty(trustStoreLocation)) {
-                    // if provided trust store is different use it and mark as config is need to changed
-                    if (!trustStoreLocation.equalsIgnoreCase(configuredTrustStore)) {
-                        isOverwriteRequired = true;
-                    }
-                    // if trust store is not provided, get the previous trust store
-                } else {
-                    if (StringUtils.isEmpty(configuredTrustStore)) {
-                        isOverwriteRequired = true;
-                        if ((trustStoreLocation = promptForTextInput(
-                                "Enter Trust store location: [" + RESTServiceConstants.DEFAULT_TRUSTSTORE_PATH + "]"))
-                                .trim().isEmpty()) {
-                            trustStoreLocation = RESTServiceConstants.DEFAULT_TRUSTSTORE_PATH;
-                        }
-                    } else {
-                        trustStoreLocation = configuredTrustStore;
-                    }
-                }
-            } else {
-                // If there is no config file, prompt user to enter
-                if (StringUtils.isEmpty(trustStoreLocation)) {
-                    isOverwriteRequired = true;
-                    if ((trustStoreLocation = promptForTextInput(
-                            "Enter Trust store location: [" + RESTServiceConstants.DEFAULT_TRUSTSTORE_PATH + "]"))
-                            .trim().isEmpty()) {
-                        trustStoreLocation = RESTServiceConstants.DEFAULT_TRUSTSTORE_PATH;
-                    }
-                }
-            }
-
-            //configure trust store password
-            //If config is there
-            if (config != null) {
-                //get previous trust store password
-                String encryptedPass = config.getToken().getTrustStorePassword();
-                String configuredTrustStorePass;
-                if(StringUtils.isEmpty(encryptedPass)){
-                    configuredTrustStorePass = null;
-                } else {
-                    try {
-                        configuredTrustStorePass = GatewayCmdUtils.decrypt(encryptedPass, new String(password));
-                    } catch (CliLauncherException e) {
-                        //different password used to encrypt
-                        configuredTrustStorePass = null;
-                    }
-                }
-
-                // check trust store password is also provided
-                if (!StringUtils.isEmpty(trustStorePassword)) {
-                    // if provided trust store password is different use it and mark as config is need to changed
-                    if (!trustStorePassword.equalsIgnoreCase(configuredTrustStorePass)) {
-                        isOverwriteRequired = true;
-                    }
-                    // if trust store password is not provided, get the previous trust store password
-                } else {
-                    if (StringUtils.isEmpty(configuredTrustStorePass)) {
-                        isOverwriteRequired = true;
-                        if ((trustStorePassword = promptForTextInput("Enter Trust store password: [ use default? ]"))
-                                .trim().isEmpty()) {
-                            trustStorePassword = RESTServiceConstants.DEFAULT_TRUSTSTORE_PASS;
-                        }
-                    } else {
-                        trustStorePassword = configuredTrustStorePass;
-                    }
-                }
-            } else {
-                // If there is no config file, prompt user to enter
-                if (StringUtils.isEmpty(trustStorePassword)) {
-                    isOverwriteRequired = true;
-                    if ((trustStorePassword = promptForTextInput("Enter Trust store password: [ use default? ]")).trim()
-                            .isEmpty()) {
-                        trustStorePassword = RESTServiceConstants.DEFAULT_TRUSTSTORE_PASS;
-                    }
-                }
-            }
-
-            File trustStoreFile = new File(trustStoreLocation);
-            if (!trustStoreFile.isAbsolute()) {
-                trustStoreLocation = GatewayCmdUtils.getCLIHome() + File.separator + trustStoreLocation;
-            }
-            trustStoreFile = new File(trustStoreLocation);
-            if (!trustStoreFile.exists()) {
-                System.err.println("Error while loading trust store location: " + trustStoreLocation);
-                Runtime.getRuntime().exit(1);
-            }
-
-            //set the trustStore
-            System.setProperty("javax.net.ssl.keyStoreType", "pkcs12");
-            System.setProperty("javax.net.ssl.trustStore", trustStoreLocation);
-            System.setProperty("javax.net.ssl.trustStorePassword", trustStorePassword);
-
-            OAuthService manager = new OAuthServiceImpl();
-            if (config != null) {
-                clientID = config.getToken().getClientId();
-                String encryptedSecret = config.getToken().getClientSecret();
-                if (!StringUtils.isEmpty(clientID.trim()) && !StringUtils.isEmpty(encryptedSecret.trim())) {
-                    try {
-                        clientSecret = GatewayCmdUtils.decrypt(encryptedSecret, new String(password));
-                    } catch (CliLauncherException e) {
-                        //different password used to encrypt
-                        clientSecret = null;
-                    }
-                }
-            }
-
-            if (StringUtils.isEmpty(clientID) || StringUtils.isEmpty(clientSecret)) {
-                String[] clientInfo = manager
-                        .generateClientIdAndSecret(registrationEndpoint, username, password.toCharArray());
-                clientID = clientInfo[0];
-                clientSecret = clientInfo[1];
-            }
-
-            String accessToken = manager
-                    .generateAccessToken(tokenEndpoint, username, password.toCharArray(), clientID, clientSecret);
-
-            RESTAPIService service = new RESTAPIServiceImpl(publisherEndpoint, adminEndpoint);
-            List<ExtendedAPI> apis = service.getAPIs(label, accessToken);
-            List<ApplicationThrottlePolicyDTO> applicationPolicies = service.getApplicationPolicies(accessToken);
-            List<SubscriptionThrottlePolicyDTO> subscriptionPolicies = service.getSubscriptionPolicies(accessToken);
-
-            ThrottlePolicyGenerator policyGenerator = new ThrottlePolicyGenerator();
-            CodeGenerator codeGenerator = new CodeGenerator();
-            boolean changesDetected = false;
-            try {
-                policyGenerator.generate(
-                        GatewayCmdUtils.getLabelSrcDirectoryPath(projectRoot, projectName) + File.separator
-                                + GatewayCliConstants.POLICY_DIR, applicationPolicies, subscriptionPolicies);
-                codeGenerator.generate(projectRoot, projectName, apis, true);
-                //Initializing the ballerina label project and creating .bal folder.
-                InitHandler.initialize(Paths.get(GatewayCmdUtils.getLabelDirectoryPath(projectRoot, projectName)), null,
-                        new ArrayList<>(), null);
-                try {
-                    changesDetected = HashUtils.detectChanges(apis, subscriptionPolicies, applicationPolicies);
-                } catch (HashingException e) {
-                    outStream.println("Error while checking for changes of resources. Skipping no-change detection..");
-                    Runtime.getRuntime().exit(1);
-                }
-            } catch (IOException | BallerinaServiceGenException e) {
-                outStream.println("Error while generating ballerina source");
-                e.printStackTrace();
-                Runtime.getRuntime().exit(1);
-            }
-
-            //if all the operations are success, write new config to file
-            if (isOverwriteRequired) {
-                Config newConfig = new Config();
-                Client client = new Client();
-                client.setHttpRequestTimeout(1000000);
-                newConfig.setClient(client);
-                
-                String encryptedSecret = GatewayCmdUtils.encrypt(clientSecret, new String(password));
-                String encryptedTrustStorePass = GatewayCmdUtils.encrypt(trustStorePassword, new String(password));
-                Token token = new TokenBuilder()
-                        .setPublisherEndpoint(publisherEndpoint)
-                        .setAdminEndpoint(adminEndpoint)
-                        .setRegistrationEndpoint(registrationEndpoint)
-                        .setTokenEndpoint(tokenEndpoint)
-                        .setUsername(username)
-                        .setClientId(clientID)
-                        .setClientSecret(encryptedSecret)
-                        .setTrustStoreLocation(trustStoreLocation)
-                        .setTrustStorePassword(encryptedTrustStorePass)
-                        .build();
-                newConfig.setToken(token);
-                newConfig.setCorsConfiguration(GatewayCmdUtils.getDefaultCorsConfig());
-                GatewayCmdUtils.saveConfig(newConfig, configPath);
-
-                //There should not be any logic after this system exit
-                if (!changesDetected) {
-                    outStream.println("No changes from server.");
-                    Runtime.getRuntime().exit(GatewayCliConstants.EXIT_CODE_NOT_MODIFIED);
-                }
-            }
-        }
-
-        @Override
-        public String getName() {
-            return GatewayCliCommands.HELP;
-        }
-
-        @Override
-        public void setParentCmdParser(JCommander parentCmdParser) {
-        }
-
-        private String promptForTextInput(String msg) {
-            outStream.println(msg);
-            return System.console().readLine();
-        }
-
-        private String promptForPasswordInput(String msg) {
-            outStream.println(msg);
-            return new String(System.console().readPassword());
-        }
-
-        private void populateHosts(String host) {
-            try {
-                publisherEndpoint = new URL(new URL(host), RESTServiceConstants.PUB_RESOURCE_PATH).toString();
-                adminEndpoint = new URL(new URL(host), RESTServiceConstants.ADMIN_RESOURCE_PATH).toString();
-                registrationEndpoint = new URL(new URL(host), RESTServiceConstants.DCR_RESOURCE_PATH).toString();
-                tokenEndpoint = new URL(new URL(host), RESTServiceConstants.TOKEN_PATH).toString();
-            } catch (MalformedURLException e) {
-                throw new RuntimeException("Malformed URL is provided");
-            }
-        }
-
-    }
-
-    /**
-     * This class represents the "build" command and it holds arguments and flags specified by the user.
-     */
-    @Parameters(commandNames = "build", commandDescription = "micro gateway build information")
-    private static class BuildCmd implements GatewayLauncherCmd {
-
-        @Parameter(names = "--java.debug", hidden = true)
-        private String javaDebugPort;
-
-        @Parameter(names = { "-n", "--project" }, hidden = true)
-        private String projectName;
-
-        @Parameter(names = { "--help", "-h", "?" }, hidden = true, description = "for more information")
-        private boolean helpFlag;
-
-        @Parameter(arity = 1)
-        private List<String> argList;
-
-        private JCommander parentCmdParser;
-
-        public void execute() {
-            if (helpFlag) {
-                String commandUsageInfo = GatewayLauncherCmd.getCommandUsageInfo("build");
-                outStream.println(commandUsageInfo);
-                return;
-            }
-
-            if (StringUtils.isEmpty(projectName)) {
-                outStream.println("Label can't be empty. " + "You need to specify -n <project name>");
-                return;
-            }
-
-            try {
-                String projectRoot = GatewayCmdUtils.getStoredWorkspaceLocation();
-                GatewayCmdUtils.createLabelGWDistribution(projectRoot, projectName);
-            } catch (IOException e) {
-                outStream.println(
-                        "Error while creating micro gateway distribution for project " + projectName + ". Reason: " + e
-                                .getMessage());
-                Runtime.getRuntime().exit(1);
-            }
-            Runtime.getRuntime().exit(0);
-        }
-
-        @Override
-        public String getName() {
-            return GatewayCliCommands.BUILD;
-        }
-
-        @Override
-        public void setParentCmdParser(JCommander parentCmdParser) {
-            this.parentCmdParser = parentCmdParser;
-        }
-    }
-
-    /**
-     * This class represents the "run" command and it holds arguments and flags specified by the user.
-     */
-    @Parameters(commandNames = "run", commandDescription = "micro gateway run information")
-    private static class RunCmd implements GatewayLauncherCmd {
-
-        @Parameter(names = "--java.debug", hidden = true)
-        private String javaDebugPort;
-
-        @Parameter(names = { "-l", "--label" }, hidden = true)
-        private String label;
-
-        @Parameter(names = { "--help", "-h", "?" }, hidden = true, description = "for more information")
-        private boolean helpFlag;
-
-        @Parameter(arity = 1)
-        private List<String> argList;
-
-        private JCommander parentCmdParser;
-
-        public void execute() {
-            if (helpFlag) {
-                String commandUsageInfo = GatewayLauncherCmd.getCommandUsageInfo("run");
-                outStream.println(commandUsageInfo);
-                return;
-            }
-
-            if (StringUtils.isEmpty(label)) {
-                outStream.println("Label can't be empty. " + "You need to specify -l <label name>");
-                return;
-            }
-        }
-
-        @Override
-        public String getName() {
-            return GatewayCliCommands.RUN;
-        }
-
-        @Override
-        public void setParentCmdParser(JCommander parentCmdParser) {
-            this.parentCmdParser = parentCmdParser;
-        }
-    }
-
-    /**
-     * This class represents the "main" command required by the JCommander.
-     */
-    private static class DefaultCmd implements GatewayLauncherCmd {
-
-        @Parameter(names = { "--help", "-h", "?" }, hidden = true, description = "for more information")
-        private boolean helpFlag;
-
-        @Parameter(names = "--java.debug", hidden = true)
-        private String javaDebugPort;
-
-        @Override
-        public void execute() {
-            if (helpFlag) {
-                printUsageInfo(GatewayCliCommands.HELP);
-                return;
-            }
-            printUsageInfo(GatewayCliCommands.DEFAULT);
-        }
-
-        @Override
-        public String getName() {
-            return GatewayCliCommands.DEFAULT;
-        }
-
-        @Override
-        public void setParentCmdParser(JCommander parentCmdParser) {
-        }
-    }
 }
